Removing dead actors and dead e2e code
The actors request cert and upgrade are no longer used. Delete those and
moved two funcs from upgrade to partitioned_update.

The e2e testdata is no longer used because we cannot run the e2e tests
against multiple different types of platform with the way those e2e
tests were written.
