+++ pkg/actor/BUILD.bazel
@@ -11,9 +11,7 @@ go_library(
         "generate_cert.go",
         "initialize.go",
         "partitioned_update.go",
-        "request_cert.go",
         "resize_pvc.go",
-        "upgrade.go",
         "validate_version.go",
     ],
     importpath = "github.com/cockroachdb/cockroach-operator/pkg/actor",
@@ -29,7 +27,6 @@ go_library(
         "//pkg/resource:go_default_library",
         "//pkg/scale:go_default_library",
         "//pkg/security:go_default_library",
-        "//pkg/tls:go_default_library",
         "//pkg/update:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/utilfeature:go_default_library",
@@ -40,7 +37,6 @@ go_library(
         "@com_github_pkg_errors//:go_default_library",
         "@io_k8s_api//apps/v1:go_default_library",
         "@io_k8s_api//batch/v1:go_default_library",
-        "@io_k8s_api//certificates/v1beta1:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
